add function para carregar foto de perfil
