+++ AdCars/Views/MyProfileView.xaml.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿using AdCars.Services;
+using ImageToArray;
+using Plugin.Media;
+using Plugin.Media.Abstractions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -12,9 +16,49 @@ namespace AdCars.Views
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class MyProfileView : ContentPage
     {
+        private MediaFile file;
         public MyProfileView()
         {
             InitializeComponent();
+        }
+
+
+        public async void PegarImagemGaleria()
+        {
+            await CrossMedia.Current.Initialize();
+
+            if (!CrossMedia.Current.IsCameraAvailable || !CrossMedia.Current.IsTakePhotoSupported)
+            {
+                await Application.Current.MainPage.DisplayAlert("ERRO", "Desculpe, seu dispositivo não suporta essa funcionalidade ", "OK");
+                return;
+            }
+
+            file = await CrossMedia.Current.PickPhotoAsync();
+
+            if (file == null)
+                return;
+
+            ProfileImage.Source = ImageSource.FromStream(() =>
+            {
+                var stream = file.GetStream();
+                AddImageToServer();
+                return stream;
+            });
+        }
+        //metodo para adicionar a imagem para o server
+        private async void AddImageToServer()
+        {
+            var ImageArray = FromFile.ToArray(file.GetStream());
+            file.Dispose();
+
+            var resposta = await ApiService.EditarFotoPerfil(ImageArray);
+            if (resposta) return;
+            await DisplayAlert("Algo deu errado", "Por favor envie a foto novamente", "OK");
+        }
+
+        private void TapGestureRecognizer_Tapped(object sender, EventArgs e)
+        {
+            PegarImagemGaleria();
         }
     }
 }
